profiles/requirements.txt
psycopg2-binary
pandas
PyYAML